refactor: split site rendering by responsibility
Break the static-site rendering helpers into focused template, conversation-page, and aggregate-index modules while keeping polylogue.site.rendering as the stable facade.
